scripts/viper_report.py- making data_uri snakemake version agnostic
NOTE: starting in snakemake 5.1.0, data_uri returns a tuple (uri, mime)
rather than just uri (which is the only thing this reporting script requires).

SOLN: we added the viper_data_uri fn to help be more independent of the
snakemake version that is used.  whether data_uri returns a tuple or not,
viper_data_uri will always return a uri